stock_analyzer.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Stock Analyzer - Professional stock analysis for A-shares, HK, and US markets
"""
import os
import sys
import argparse
import json
from datetime import datetime
from typing import Optional, Dict, Any
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class StockAnalyzer:
    """Simplified stock analyzer for OpenClaw skill"""

    def __init__(self):
        self.api_key = os.getenv('MINIMAX_API_KEY') or os.getenv('OPENAI_API_KEY')
        if not self.api_key:
            print("⚠️  Warning: No API key found. Set MINIMAX_API_KEY or OPENAI_API_KEY in .env")
            sys.exit(1)

    def analyze(self, code: str, detailed: bool = False) -> Dict[str, Any]:
        """
        Analyze a stock and return structured result

        Args:
            code: Stock code (e.g., 600519, 00700, AAPL)
            detailed: Whether to generate detailed report

        Returns:
            Dictionary with analysis results
        """
        # TODO: Integrate with existing analysis system
        # For now, return mock data to demonstrate structure

        result = {
            'code': code,
            'name': self._get_stock_name(code),
            'score': self._calculate_score(code),
            'recommendation': self._get_recommendation(code),
            'trend': self._get_trend(code),
            'timestamp': datetime.now().isoformat()
        }

        if detailed:
            result.update({
                'technical_indicators': self._get_technical_indicators(code),
                'ai_insights': self._get_ai_insights(code),
                'support_resistance': self._get_support_resistance(code)
            })

        return result

    def _get_stock_name(self, code: str) -> str:
        """Get stock name from code"""
        # Simple mapping - TODO: Integrate with data sources
        names = {
            '600519': '贵州茅台',
            '00700': '腾讯控股',
            'AAPL': 'Apple Inc.',
            'TSLA': 'Tesla Inc.',
            'NVDA': 'NVIDIA Corp.'
        }
        return names.get(code, 'Unknown')

    def _calculate_score(self, code: str) -> int:
        """Calculate analysis score (0-100)"""
        # TODO: Implement actual scoring logic
        # For demo, return different scores based on code
        scores = {'600519': 72, '00700': 65, 'AAPL': 75, 'TSLA': 45, 'NVDA': 78}
        return scores.get(code, 50)

    def _get_recommendation(self, code: str) -> str:
        """Get buy/hold/sell recommendation"""
        score = self._calculate_score(code)
        if score >= 70:
            return '买入/持有'
        elif score >= 50:
            return '持有/观望'
        else:
            return '观望/卖出'

    def _get_trend(self, code: str) -> str:
        """Get trend information"""
        # TODO: Implement actual trend analysis
        return '多头排列 ↑' if self._calculate_score(code) >= 60 else '空头排列 ↓'

    def _get_technical_indicators(self, code: str) -> Dict[str, Any]:
        """Get technical indicators"""
        # TODO: Implement actual technical analysis
        return {
            'ma5': 1680.5,
            'ma10': 1675.2,
            'ma20': 1668.8,
            'volume_change': '+15%',
            'momentum': '中性偏多'
        }

    def _get_ai_insights(self, code: str) -> str:
        """Get AI-powered insights"""
        # TODO: Integrate with AI analysis
        insights = {
            '600519': '作为高端白酒龙头，茅台护城河深厚。近期北向资金持续流入。',
            '00700': '低估值互联网龙头，持续回购。关注游戏业务恢复情况。',
            'AAPL': '硬件销量稳健，服务业务增长强劲。估值合理。',
            'TSLA': '电动车市场竞争加剧，关注自动驾驶进展和储能业务。',
            'NVDA': 'AI基础设施核心受益者，数据中心需求持续旺盛。'
        }
        return insights.get(code, '暂无AI分析')

    def _get_support_resistance(self, code: str) -> Dict[str, Any]:
        """Get support and resistance levels"""
        # TODO: Implement actual calculation
        return {
            'support': [1675, 1650],
            'resistance': [1700, 1725]
        }

    def format_output(self, result: Dict[str, Any], detailed: bool = False, output_format: str = 'text') -> str:
        """Format analysis result for output"""
        if output_format == 'json':
            return json.dumps(result, ensure_ascii=False, indent=2)

        # Text format
        lines = [
            f"{result['code']} {result['name']}",
            "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━",
            f"评分: {result['score']}/100",
            f"建议: {result['recommendation']}",
            f"趋势: {result['trend']}"
        ]

        if detailed:
            lines.append("\n技术指标:")
            ti = result.get('technical_indicators', {})
            lines.append(f"  MA趋势: MA5({ti.get('ma5', 'N/A')}) > MA10({ti.get('ma10', 'N/A')}) > MA20({ti.get('ma20', 'N/A')})")
            lines.append(f"  成交量: {ti.get('volume_change', 'N/A')}")
            lines.append(f"  动量: {ti.get('momentum', 'N/A')}")

            sr = result.get('support_resistance', {})
            lines.append(f"\n支撑/压力:")
            lines.append(f"  支撑: {', '.join(map(str, sr.get('support', [])))}")
            lines.append(f"  压力: {', '.join(map(str, sr.get('resistance', [])))}")

            lines.append(f"\n💡 AI 洞察:")
            lines.append(f"  {result.get('ai_insights', '暂无')}")

        return '\n'.join(lines)


def main():
    """Main entry point"""
    parser = argparse.ArgumentParser(
        description='Stock Analyzer - Professional stock analysis tool',
        formatter_class=argparse.RawDescriptionHelpFormatter
    )

    parser.add_argument(
        'codes',
        nargs='+',
        help='Stock codes to analyze (e.g., 600519, 00700, AAPL)'
    )

    parser.add_argument(
        '--detailed', '-d',
        action='store_true',
        help='Generate detailed report'
    )

    parser.add_argument(
        '--output', '-o',
        choices=['text', 'json'],
        default='text',
        help='Output format (default: text)'
    )

    args = parser.parse_args()

    # Initialize analyzer
    analyzer = StockAnalyzer()

    # Analyze each stock
    for code in args.codes:
        result = analyzer.analyze(code, detailed=args.detailed)
        print(analyzer.format_output(result, detailed=args.detailed, output_format=args.output))

        if len(args.codes) > 1:
            print()  # Add separator between stocks


if __name__ == '__main__':
    main()
